backend/pyproject.toml
[tool.poetry]
name = "tuberry-backend"
version = "0.1.0"
description = "Tuberry MVP backend"
authors = ["Tuberry Team <dev@tuberry.local>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlmodel = "^0.0.16"
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
httpx = "^0.27.0"
redis = "^5.0.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
loguru = "^0.7.2"
bcrypt = "4.1.3"

[tool.poetry.group.dev.dependencies]
black = "^24.4"
ruff = "^0.3.5"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
